fixup! fixup! Don't exclude quite so much in a no-sock build
change unsigned long -> unsigned int and adapt socket detection in bio_wait()

+++ include/openssl/bio.h
@@ -663,7 +663,7 @@ int BIO_sock_should_retry(int i);
 int BIO_sock_non_fatal_error(int error);
 int BIO_socket_wait(int fd, int for_read, time_t max_time);
 # endif
-int BIO_wait(BIO *bio, time_t max_time, unsigned long milliseconds);
+int BIO_wait(BIO *bio, time_t max_time, unsigned int milliseconds);
 int BIO_connect_retry(BIO *bio, int timeout);
 
 int BIO_fd_should_retry(int i);
